ARM: dts: override p4note poweroff driver
the p4note device has a special feature that the device will reboot
instead of shutdown if there is a power cable attached. this driver sets
the right registers so that this charging mode is detected and accounted
for. the effect is that the bootloader will set a kernel cmd line
parameter which android handles to start charging mode instead of
booting into the os.

Signed-off-by: Martin Jücker <martin.juecker@gmail.com>